fix compiler warning:
gcc -DHAVE_CONFIG_H -I. -I./src    -DOPENBSD -DSYSCONFDIR="\"/usr/local/etc\"" -I./src -I./src/device -I./src/http -I./src/process -I./src/protocols -I./src/ssl -I./libmonit/src   -Wno-address -Wno-pointer-sign -g -O2 -Wall -Wunused -Wno-unused-label -funsigned-char -D_GNU_SOURCE -std=c99 -D _REENTRANT -I/usr/include -c -o src/protocols/generic.o src/protocols/generic.c
src/protocols/generic.c: In function 'check_generic':
src/protocols/generic.c:97: warning: comparison is always false due to limited range of data type

+++ src/protocols/generic.c
@@ -94,7 +94,7 @@ void check_generic(Socket_T socket) {
                          buffer and then set a low timeout on next read which reads remaining bytes
                          as well as wait on EOF */
                         *buf = Socket_readByte(socket);
-                        if (*buf < 0) {
+                        if ((int8_t)*buf < 0) {
                                 FREE(buf);
                                 THROW(IOException, "GENERIC: error receiving data -- %s", STRERROR);
                         }
